+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The tag [OUTPUT] indicates changes affecting the output files.
 
 - [INPUT] Add the `ExponentialUnadjusted` learning model (exponential model with no adjustment for
   the first iteration).
+- [INPUT] Add option to force the route to be taken for road legs (parameter `route`).
 - [OUTPUT] Add the root mean squared difference of departure time in the iteration results.
 - [OUTPUT] Add the root mean squared difference between simulated and expected travel time in the
   iteration results.